# secrets_broker/openapi.spec.yml
openapi: "3.0.0"

info:
  title: GitHub Actions Secrets Broker
  version: "0.1"

paths:
  /api/v1/initialize:
    get:
      summary: Initialize secrets request
      description: Get token for github action to get secrets request validated.
      operationId: secrets_broker.api.initialize
      parameters:
        - name: x-run-id
          in: header
          required: true
          schema:
            type: integer
        - name: x-github-repo
          in: header
          required: true
          schema:
            type: string
        - name: x-github-token
          in: header
          required: true
          schema:
            type: string
      responses:
        200:
          description: Validation token.
          content:
            application/json:
              schema:
                type: object
                properties:
                  validation_token:
                    type: string
                required:
                  - validation_token
        403:
          description: GitHub repo not allowed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMsg'
        409:
          description: Repo & Run ID already initialized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMsg'

  /api/v1/secrets:
    get:
      summary: Get Secrets
      description: Get list of secrets.
      operationId: secrets_broker.api.secrets
      parameters:
        - name: x-run-id
          in: header
          required: true
          schema:
            type: integer
        - name: x-github-repo
          in: header
          required: true
          schema:
            type: string
        - name: x-github-token
          in: header
          required: true
          schema:
            type: string
        - name: keys
          description: Comma-separated secret names
          in: query
          required: true
          schema:
            type: string
      responses:
        200:
          description: List of secrets.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    key:
                      type: string
                    value:
                      type: string
                  required:
                    - key
                    - value
        403:
          description: Access to secrets forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMsg'

components:
  schemas:
    ErrorMsg:
      type: object
      properties:
        error:
          type: string
      required:
        - error